pdf_chatbot/requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
sentence-transformers
pdfminer.six
google-generativeai
